Replace development server to use webpack output
The reset logic was moved to the build script, since a new server object is created after each rebuild.
A global variable 'env' was created so that the prod server is started only it's not in the watch mode.
A method to unregister the pretty log from pino was created to remove the output listener before creating a new instance.